requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
supabase
pymupdf4llm
google-genai
python-multipart
azure-search-documents
azure-core
openai